Use atomic access for shmem
Use atomic access for access xid values in shared memory.
This was coused problem on 32-bit arch.

+++ src/include/access/transam.h
@@ -338,7 +338,7 @@ ReadNextTransactionId(void)
 
 /* return transaction ID backed up by amount, handling wraparound correctly */
 static inline TransactionId
-TransactionIdRetreatedBy(TransactionId xid, uint32 amount)
+TransactionIdRetreatedBy(TransactionId xid, uint64 amount)
 {
 	xid -= amount;
